+++ folioflex/dashboard/pages/ideas.py
@@ -1,8 +1,23 @@
 """Ideas dashboard."""
 
-from dash import dash_table, dcc, html
+import dash
+import pandas as pd
+from dash import Input, Output, State, callback, dash_table, dcc, html
 
 from folioflex.dashboard.utils import dashboard_helper
+from folioflex.portfolio import wrappers
+from folioflex.utils import custom_logger
+
+logger = custom_logger.setup_logging(__name__)
+
+dash.register_page(__name__, path="/", title="folioflex - Stocks", order=0)
+
+#   _                            _
+#  | |    __ _ _   _  ___  _   _| |_
+#  | |   / _` | | | |/ _ \| | | | __|
+#  | |__| (_| | |_| | (_) | |_| | |_
+#  |_____\__,_|\__, |\___/ \__,_|\__|
+#              |___/
 
 
 def layout(login_status, login_alert):
@@ -57,3 +72,40 @@ def layout(login_status, login_alert):
             ),
         ]
     )
+
+
+#    ____      _ _ _                _
+#   / ___|__ _| | | |__   __ _  ___| | _____
+#  | |   / _` | | | '_ \ / _` |/ __| |/ / __|
+#  | |__| (_| | | | |_) | (_| | (__|   <\__ \
+#   \____\__,_|_|_|_.__/ \__,_|\___|_|\_\___/
+
+
+@callback(
+    [
+        Output(component_id="sma-table", component_property="columns"),
+        Output(component_id="sma-table", component_property="data"),
+    ],
+    [Input(component_id="sma-button", component_property="n_clicks")],
+    [State(component_id="idea-input", component_property="value")],
+)
+def sma_value(n_clicks, input_value):
+    """Provide simple moving average on ideas."""
+    if n_clicks == 0:
+        sma_table = (None, None)
+    else:
+        sma = wrappers.Yahoo().get_sma(ticker=input_value, days=365)
+        latest_price = wrappers.Yahoo().fast_info(ticker=input_value)["lastPrice"]
+        change_percent = wrappers.Yahoo().get_change_percent(
+            ticker=input_value, days=365
+        )
+
+        # build table
+        sma_table = [input_value, sma, latest_price, change_percent]
+        df = pd.DataFrame(
+            [sma_table], columns=["stock", "sma", "latest_price", "change_percent"]
+        )
+
+        sma_table = [{"name": i, "id": i} for i in df.columns], df.to_dict("records")
+
+    return sma_table
