+++ pyproject.toml
@@ -28,6 +28,7 @@ dependencies = [
 
     # web
     "dash>=1.0.2",
+    "dash_bootstrap_components>=1.6.0",
     "dash-core-components>=1.0.0",
     "dash-html-components>=1.0.0",
     "dash-renderer>=1.0.0",
